requirements.txt
beautifulsoup4==4.11.1
fastapi==0.85.0
psycopg2==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.26.0
SQLAlchemy==1.4.41
uvicorn==0.12.3
aiofiles==0.6.0
